Some more work on the UI.
Changes are mostly to have a cleaner set of posting pages.
The form for posting text messages or photos are now inside
a div which is not using all the width and has some padding.
Also the entry have now some padding to avoid the text to be
close to the borders.