New trigger to set last_comment_id for new post.
We set the last_comment_id to current value plus one. This
is to have proper sorting for Dernièrement commentées. The
idea is to get new post listed into the flow when using this
sort option as it is more user friendly. So it means that
last_comment_id is meaningful only when comment_counter is
not zero.