Fix handling of context in the test suite.
The test suite was failing as the Web_Block support now
properly handle the unknown context token. So we want to
make sure the context is always passed to the URL. At the
same time the code to get the context out of the Web page
has now been factorized.