Add information about the CdC in post page.
The coup-de-coeur are now tagged into the post page and the
vote button is replaced by a text saying that this post has
already been selected as CdC as we do not want to let user
vote for a photo that has already been a CdC in the past.
The red-heart is a link that will go to the CdC page later.