Move all forum filters HTML select into its own file.
All HTML widget are now into a chunk file as it will be
reused for the user's preference settings.
Note that at the same time a new file set_values.incl has been
introduced. This file contains all the select values that
are to be used on the HTML and Ada side.