The user preferences must not be set to the current settings.
By reusing the callbacks for setting the default values for the
select widgets the defaults values displayed were in fact the
current values set for navigating the threads. The callbacks
for the current settings and for the user preferences are now
distinct.
At the same time the user preferences have all been grouped
into a record for easy manipulation (especially since we will
be adding some more settings like the default image size
later).
--
Finale note about the user's preferences implementation is that
the filter is now set to "seven last days" as a default instead
of "all messages". This seems a good default and will stress
less the database.