Introduce a Web block for displaying comments.
The comments were handled as part of the page forum entry.
Comments are now displayed using a block (lazy tag). In the
database the forum entry information and set of comments have
been separated into two routines. The Ajax response can now
refresh all the comments and not only adding the last comment
entered by the user. It means that other comments added while
commenting a post are displayed.

Note that all the support for threaded comments in the Ajax
reply has been removed as this won't be needed anymore. All
the comments are read back from the database and sent to the
user's browser. This will take a bit more bandwidth (but text
are not that big) but will ease a lot the threaded implementation
and is needed anyway to have proper refresh.