Optimize handling of navigations links.
To compute the navigation link all the threads information
was retrieve but only the thread ids are needed. The code
to handle this was duplicated. The handling of threads
and navigation are now in a single place (Navigation_Links).
Note that setting the navigation has been removed from the
spec as all is handled internally to this unit. This is
cleaner and will easy maintenance.