Add date/time of forum last activity.
This information is displayed next to the forum name on the
top of the screen. Only the time is dispalyed if the last
message is from the current day. The date is displayed otherwise.
This is to avoid using too much space. Note that the last
activity timestamp is computed using triggers.