Set From_Main for links where context must be generated.
When navigating from a direct link (instead of navigating
from the forum next/previous) we need to build the context
like the current forum id. This is true not only for the
links from the lasted post on main page but also for the
CdC on the main page and on the CdC page.
--
Fixes display of the category chooser in entry page when
login-in after using a CdC post link.