Better implementation for detecting new context need.
The need for new context was based on the history length.
First this does not work properly across different browser.
On IE the initial history length is 0, on Firefox and
Safari it is 1.
The new implementation is based on the referer. The referer
is empty when entering an URL directly into the browser.
This is exactly what the AWS Ajax kernel want to detect.