Improve the context handling.
The previous patch was good at handling a cut&paste of an
URL into a new page. Opening a new page while clicking
on a link was not working anymore. We now check for the
presense of the referer header (handle first case) and the
history length (handle the second case).