Full rewrite of the exif support.
The previous implementation was expecting that exif parameters
was always presented into the same order. This is not right
in many cases. The new implementation is now independant of the
output ordering. At the same time some aliases for EXIF parameters
have been introduced (not all information are presented with the
same name).

Add corresponding test.