Add information about last modification instead of date of post.
It is more user fiendly to output the date of last modification for
the text and photo based forums. On text based forums we get the
post date as a tooltip over the last modification date. For photo
based forum we now output this date after the counters and the
post date as a tooltip over the thumbnail.