Clear TID in context when posting a new photo.
This is important as we are going to post a new photo. We do not
want to keep track of the previously visited photo ortherwise
some information about the photo are passed to the post template.
It was the case for the geo metadata, confusing the post page
which was displaying the location of the new photo instead of
proposing to locate it.