All libraries must be dynamic.
The ImageMagick and Jpeg libraries were set to static as
this was working find on UNIX and on Windows we had import
libraries. We now set those libraries to dynamic as it works
also on Windows (-lmagic_wand will pick the import library)
and fixes build issue with recent GNAT versions.