Add new config parameter Default_Timezone.
The default value is set to +1 (GMT + 1 is french timezone). So
when no timezone is present we will use the default timezone.

The default value for the default timezone is set to the french
timezone as this will work fine most of the time for a french
target audience like v2p.