Add support for handling client's timezone.
The different dates and times displayed on the Web page were all
local time of the server. This was confusing. To report all times
in client's local time we use the following scheme.

1. The Javascript AWS runtime will compute and set the timezone
   offset to the web browser using an Ajax request.
   This is done only if no timezone (div whose id is CTX_TZ) present
   on the web page.

2. The server s modified to handle the Ajax request and set the timezone
   in the session and in the context.

3. All dates and times from the database (stored as UTC timestamp) are
   returned using this timezone.

This works fine except for the initial page of a new Web session as
the timezone is not known at this point. If this proves to be a
problem it should be possible to send an initial request to get
the timezone and then redirect to the requested page.