Rework the SQL statement to compute the user's comment list.
This new version is lot faster than the previous one.

On on user where it has been tested it was running 11 secondes where
it take now 0.10 secondes.

At the same time the vote comment have been removed from the query. It
was anyway removed on the template. It is better to remove them during
the select.