Add a timestamp to the remember_user table.
This is needed to be able to clean-up this table. A user without
cookie support or removing cookie regularly will have a new entry
created each time he connect to the server. At some point we will
need to remove old entries. This timestamp is there to be able to
do that.

Note that the migration script copy the database as SQLite does
not support non-constant default values.