Fix testsuite to pass with the new context support.
Now that the context id changes for every request we need
to retrieve it every time. For this wrappers (Call,Upload)
have been created that send the request to the server and
record the current context internally.

Note that with this new context support a problem in the
testsuite has been detected. In fact, a test was changing
the filters after loading the root web page. The filters
are not available from this page and FID was not set (but
was needed by the framework).