Fix handling of UTF-8 named.
The templates parser SLICE filter is not UTF-8 compliant. The cut is
now done using the CSS.

This problem was preventing Ajax actions to properly work on Windows
(Internet Explorer and Firefox). This is because the same XML parser
is used on this platform. A wrong cut in an UTF-8 name was producing
an invalid UTF-8 XML response which was rejected by the XML parser.