Add user login into the lost password message.
Only the password was sent, but it can be the case that the
login is not remembered too. Seems cleaner to send the login
and the password in this case.

The Get_Password_From_Email routine has been replaced by the
more general Get_User_Data_From_Email for this purpose.