Add support for collapsing comments.
This adds a new feature (controlled by a preference) where it is
possible to collapse comments. It is then easier to comment without
being influenced by previous comments. Comments are always available
as only collapse.

The default is comments visible, so this does not change the current
UI and user experience. Users can control permanently the behavior
by setting the corresponding preference in the user page.