CdC alert is now based on proper score computation.
Using the code developped for displaying the CdC score it is now possible
to compute the current CdC score and then have an accurate alert when the
score is not ok for the next election. Note that the score is cached and
only recomputed when needed. This avoid putting to much pressure on
the database engine.